requirements.txt
streamlit==1.24.0
openai==0.27.8
pandas==2.0.3
numpy==1.24.3
Pillow==9.5.0
python-dotenv==1.0.0
bcrypt==4.0.1
streamlit-ace==0.1.1 